requirements.txt
pandas~=1.5.1
numpy~=1.23.4
tqdm
sklearn
yellowbrick
xgboost
streamlit~=1.14.1
Pillow~=9.3.0